fix: [AICPS-55] Refactor naming convention of modal components
