+++ src/components/UserPage/WelcomeSection.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 
-import ConfirmBeforeTestModal from '../LandingPage/ConfirmBeforeTestModal';
+import ConfirmBeforeTestModal from '../../pages/LandingPage/ConfirmBeforeTestModal';
 import StepModal from '../molecules/StepModal/StepModal';
 import { InnovationPana } from '@app/assets/images';
 import { TextTyping } from '@app/components/atoms/';
