+++ src/pages/LandingPage/PartnerNetwork.tsx
