+++ src/pages/LandingPage/homepage.scss
@@ -1,0 +1,115 @@
+@keyframes slideInFromLeft {
+  from {
+    transform: translateX(-100%);
+    opacity: 0;
+  }
+  to {
+    transform: translateX(0);
+    opacity: 1;
+  }
+}
+
+.slide-in-left {
+  animation: slideInFromLeft 0.8s ease-out forwards;
+}
+@keyframes fadeInText {
+  0% {
+    opacity: 0;
+    transform: translateY(10px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.fade-in-text {
+  opacity: 0;
+  animation: fadeInText 3s ease-out forwards;
+}
+@keyframes slideUp {
+  0% {
+    opacity: 0;
+    transform: translateY(600px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.slide-up {
+  opacity: 0;
+  animation: slideUp 1.5s ease-out forwards;
+}
+@keyframes zoomIn {
+  0% {
+    transform: scale(0);
+    opacity: 0;
+  }
+  100% {
+    transform: scale(1);
+    opacity: 1;
+  }
+}
+
+.bg-zoom-in {
+  animation: zoomIn 1s ease-out forwards;
+  transform-origin: center center;
+}
+@keyframes fadeInUp {
+  0% {
+    opacity: 0;
+    transform: translateY(50px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.scroll-animate {
+  opacity: 0;
+  transform: translateY(50px);
+  transition: opacity 0.3s ease-out, transform 0.3s ease-out;
+
+  &.in-view {
+    animation: fadeInUp 0.8s ease-out forwards;
+  }
+}
+@keyframes slideUpThenLeft {
+  0% {
+    opacity: 0;
+    transform: translateY(200px) translateX(200px);
+  }
+
+  50% {
+    opacity: 0.5;
+    transform: translateY(0) translateX(200px);
+  }
+
+  100% {
+    transform: translateX(0);
+    opacity: 1;
+  }
+}
+
+.slide-up-then-left {
+  animation: slideUpThenLeft 1.5s ease-in-out forwards;
+}
+.custom-banner {
+  .ant-image {
+    height: 70%;
+  }
+  .ant-image-img {
+    height: 100%;
+  }
+}
+.custom-banner-user {
+  .ant-image {
+    height: 80% !important;
+  }
+  .ant-image-img {
+    height: 80% !important;
+  }
+}
